refactor(service-portal): refactoring sidebar code  (#7710)
* refactor: styling in sidebar

* refactor: sidebar styles

* fix: hover states

* fix: remove unused import
